requirements.txt
torch==2.5.1+cu121
torchvision==0.20.1+cu121
matplotlib
numpy
Pillow
tqdm
gradio
kagglehub